Changed return ints to return longs
Those functions are returning memory addresses, and while they may work now
it may be best if they are pointers as of now even though i havnt seen anything bad happen

+++ linuxdoom-1.10/w_wad.c
@@ -379,7 +379,7 @@ int W_NumLumps (void)
 // Returns -1 if name not found.
 //
 
-int W_CheckNumForName (char* name)
+long W_CheckNumForName (char* name)
 {
     union {
 	char	s[9];
@@ -441,7 +441,7 @@ int W_CheckNumForName (char* name)
 // Accessor for lump by using the name , but errors out if its not found
 // Calls W_CheckNumForName, but bombs out if not found.
 //
-int W_GetNumForName (char* name)
+long W_GetNumForName (char* name)
 {
     int	i;
